+++ core/src/com/teamonehundred/pixelboat/ObstacleBranch.java
@@ -23,7 +23,7 @@ public class ObstacleBranch extends Obstacle {
      * @author James Frost
      * @author William Walton
      */
-    public ObstacleBranch(int x, int y) {
+    public ObstacleBranch(float x, float y) {
         super(x, y, 60, 60, "obstacle_branch.png");
         getSprite().setRotation(-90 + (float) Math.random() * 180);
     }
